src/requirements.txt
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
exceptiongroup==1.2.0
Faker==20.1.0
flasgger==0.9.7.1
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
gunicorn==20.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mistune==2.0.5
mysqlclient==2.1.1
packaging==23.1
parameterized==0.9.0
pluggy==1.3.0
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.1.0
pyrsistent==0.19.3
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
redis==5.0.1
six==1.16.0
SQLAlchemy==2.0.23
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.3.4
